01-install-eks-cluster/requirements.txt
aws-cdk.core==1.51.0
aws-cdk.aws-ec2==1.51.0
aws-cdk.aws-eks==1.51.0